requirements.txt
# Core Requirements
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1

# Security
bcrypt==4.1.2

# Database
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Forms
WTForms==3.1.1
Flask-WTF==1.2.1

# Optional but recommended
gunicorn==21.2.0
Flask-Login==0.6.3
Flask-Cors==4.0.0